converting portions of components to client side such as SearchBar and ReservationCard
