+++ src/app/restaurant/[slug]/components/Rating.tsx
