+++ src/app/restaurant/[slug]/components/ReservationCard.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 const ReservationCard = () => {
   return (
     <div className="w-[27%] relative text-reg text-black">
