+++ src/app/restaurant/[slug]/page.tsx
@@ -2,7 +2,7 @@ import NavBar from '@/app/components/NavBar';
 import Header from './components/Header';
 import RestaurantNavBar from './components/RestaurantNavBar';
 import Title from './components/Title';
-import Rating from './components/Ratings';
+import Rating from './components/Rating';
 import Description from './components/Description';
 import Images from './components/Images';
 import Reviews from './components/Reviews';
